requirements.txt
python-dotenv
supabase
reportlab
streamlit